# [CUB} Fix `DeviceScan` to avoid passing invalid elements to the scan op