--- a/cub/cub/agent/agent_scan.cuh
+++ b/cub/cub/agent/agent_scan.cuh
@@ -243,7 +243,7 @@ struct AgentScan
   // Block scan utility methods
   //---------------------------------------------------------------------
 
-  template <bool Inclusive = IS_INCLUSIVE, bool IS_LAST_TILE = false>
+  template <bool Inclusive = IS_INCLUSIVE, bool LastTile = false>
   _CCCL_DEVICE _CCCL_FORCEINLINE void ScanFirstTile(
     AccumT (&items)[ITEMS_PER_THREAD],
     InitValueT init_value,
@@ -256,7 +256,7 @@ struct AgentScan
     {
       if constexpr (HAS_INIT)
       {
-        if constexpr (IS_LAST_TILE)
+        if constexpr (LastTile)
         {
           blockScan.InclusiveScanPartialTile(
             items, items, init_value, scan_op, static_cast<int>(num_remaining), block_aggregate);
@@ -269,7 +269,7 @@ struct AgentScan
       }
       else
       {
-        if constexpr (IS_LAST_TILE)
+        if constexpr (LastTile)
         {
           blockScan.InclusiveScanPartialTile(items, items, scan_op, static_cast<int>(num_remaining), block_aggregate);
         }
@@ -281,7 +281,7 @@ struct AgentScan
     }
     else
     {
-      if constexpr (IS_LAST_TILE)
+      if constexpr (LastTile)
       {
         blockScan.ExclusiveScanPartialTile(
           items, items, init_value, scan_op, static_cast<int>(num_remaining), block_aggregate);
@@ -294,14 +294,14 @@ struct AgentScan
     }
   }
 
-  template <typename PrefixCallback, bool Inclusive = IS_INCLUSIVE, bool IS_LAST_TILE = false>
+  template <typename PrefixCallback, bool Inclusive = IS_INCLUSIVE, bool LastTile = false>
   _CCCL_DEVICE _CCCL_FORCEINLINE void ScanSubsequentTile(
     AccumT (&items)[ITEMS_PER_THREAD], ScanOpT scan_op, PrefixCallback& prefix_op, OffsetT num_remaining = TILE_ITEMS)
   {
     BlockScanT blockScan(temp_storage.scan_storage.scan);
     if constexpr (Inclusive)
     {
-      if constexpr (IS_LAST_TILE)
+      if constexpr (LastTile)
       {
         blockScan.InclusiveScanPartialTile(items, items, scan_op, static_cast<int>(num_remaining), prefix_op);
       }
@@ -312,7 +312,7 @@ struct AgentScan
     }
     else
     {
-      if constexpr (IS_LAST_TILE)
+      if constexpr (LastTile)
       {
         blockScan.ExclusiveScanPartialTile(items, items, scan_op, static_cast<int>(num_remaining), prefix_op);
       }
@@ -396,7 +396,8 @@ struct AgentScan
     {
       // Scan first tile
       AccumT block_aggregate;
-      ScanFirstTile<IS_INCLUSIVE, IS_LAST_TILE>(items, init_value, scan_op, block_aggregate, num_remaining);
+      ScanFirstTile<IS_INCLUSIVE, IS_LAST_TILE && !cub::detail::has_no_side_effects<ScanOpT, AccumT>>(
+        items, init_value, scan_op, block_aggregate, num_remaining);
 
       if ((!IS_LAST_TILE) && (threadIdx.x == 0))
       {
@@ -407,7 +408,10 @@ struct AgentScan
     {
       // Scan non-first tile
       TilePrefixCallbackOpT prefix_op(tile_state, temp_storage.scan_storage.prefix, scan_op, tile_idx);
-      ScanSubsequentTile<TilePrefixCallbackOpT, IS_INCLUSIVE, IS_LAST_TILE>(items, scan_op, prefix_op, num_remaining);
+      ScanSubsequentTile<TilePrefixCallbackOpT,
+                         IS_INCLUSIVE,
+                         IS_LAST_TILE && !cub::detail::has_no_side_effects<ScanOpT, AccumT>>(
+        items, scan_op, prefix_op, num_remaining);
     }
 
     __syncthreads();