--- a/cub/cub/agent/single_pass_scan_operators.cuh
+++ b/cub/cub/agent/single_pass_scan_operators.cuh
@@ -1268,7 +1268,8 @@ struct TilePrefixCallbackOp
     // Use the swizzled scan operator because we are now scanning *down* towards thread0.
 
     int tail_flag =
-      (predecessor_status == StatusWord(SCAN_TILE_INCLUSIVE) || predecessor_status == StatusWord(SCAN_TILE_OOB));
+      (predecessor_status == StatusWord(SCAN_TILE_INCLUSIVE)
+       || (!cub::detail::has_no_side_effects<ScanOpT, T> && predecessor_status == StatusWord(SCAN_TILE_OOB)));
     window_aggregate =
       WarpReduceT(temp_storage.warp_reduce).TailSegmentedReduce(value, tail_flag, SwizzleScanOp<ScanOpT>(scan_op));
   }