--- a/cub/cub/warp/specializations/warp_scan_shfl.cuh
+++ b/cub/cub/warp/specializations/warp_scan_shfl.cuh
@@ -412,10 +412,14 @@ struct WarpScanShfl
     _T temp = ShuffleUp<LOGICAL_WARP_THREADS>(input, offset, first_lane, member_mask);
 
     // Perform scan op if from a valid peer
-    _T output = scan_op(temp, input);
+    _T output;
     if (static_cast<int>(lane_id) < first_lane + offset)
     {
       output = input;
+    }
+    else
+    {
+      output = scan_op(temp, input);
     }
 
     return output;