… cub/test/catch2_test_block_scan_api.cu
… cub/test/catch2_test_block_scan_partial.cu
… cub/test/catch2_test_block_scan_partial_containers.cu
… cub/test/catch2_test_block_scan_partial_helper.cuh
… cub/test/catch2_test_block_scan_partial_invalid.cu
… cub/test/catch2_test_device_scan_invalid.cu