## Description

<!-- Every PR should have a corresponding issue that describes and motivates the work done in the PR -->
closes #5017

<!-- Provide a standalone description of changes in this PR. -->
Fixes all root causes identified in #5019

TODO:
- [ ] Benchmark
- [ ] Avoid overheads when both scan operator is from `cuda::std::` and type is primitive.
- [ ] Add benchmark for custom operator that needs to invoke overheads

<!-- Note: The pull request title will be included in the CHANGELOG. -->

## Checklist
- [x] I am familiar with the [Contributing Guidelines]().
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.