@pauleonix 2025-08-27
Current C Parallel/Python failures seem to be based on `LDL` in the produced SASS. I.e. this should be fixable with re-tuning.