@pauleonix 2025-08-29
Results for `cub.bench.scan.exclusive.custom.base`:

From NCU:
- old version uses local memory for
  - I128
- new version uses local memory for
  - only I128, but there it has slightly fewer local memory requests than with the old version (opposite from the other benchmark), but got slower
  - C64 but got faster

['/build/scan_custom_partile.json', '/build/scan_custom_new.json']
# base

## [0] NVIDIA H100 PCIe

|  T{ct}  |  OffsetT{ct}  |  Elements{io}  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |         Diff |   %Diff |  Status  |
|---------|---------------|----------------|------------|-------------|------------|-------------|--------------|---------|----------|
|   I8    |      I32      |      2^16      |   8.821 us |       4.69% |   8.833 us |       4.31% |     0.012 us |   0.13% |   SAME   |
|   I8    |      I32      |      2^20      |  12.692 us |       3.63% |  12.711 us |       3.11% |     0.018 us |   0.15% |   SAME   |
|   I8    |      I32      |      2^24      |  53.948 us |       3.40% |  51.932 us |       3.99% |    -2.016 us |  -3.74% |   FAST   |
|   I8    |      I32      |      2^28      | 754.029 us |       1.42% | 695.943 us |       1.39% |   -58.086 us |  -7.70% |   FAST   |
|   I8    |      I64      |      2^16      |   9.119 us |       3.90% |   9.112 us |       3.20% |    -0.007 us |  -0.07% |   SAME   |
|   I8    |      I64      |      2^20      |  12.777 us |       3.30% |  12.786 us |       3.03% |     0.009 us |   0.07% |   SAME   |
|   I8    |      I64      |      2^24      |  53.973 us |       3.36% |  52.657 us |       3.93% |    -1.316 us |  -2.44% |   SAME   |
|   I8    |      I64      |      2^28      | 752.508 us |       1.35% | 696.264 us |       1.38% |   -56.244 us |  -7.47% |   FAST   |
|   I16   |      I32      |      2^16      |   9.012 us |       4.67% |   9.071 us |       3.98% |     0.060 us |   0.67% |   SAME   |
|   I16   |      I32      |      2^20      |  13.437 us |       3.19% |  13.437 us |       2.90% |     0.000 us |   0.00% |   SAME   |
|   I16   |      I32      |      2^24      |  68.251 us |       3.94% |  65.313 us |       3.87% |    -2.938 us |  -4.30% |   FAST   |
|   I16   |      I32      |      2^28      | 948.132 us |       1.69% | 884.059 us |       1.52% |   -64.073 us |  -6.76% |   FAST   |
|   I16   |      I64      |      2^16      |   8.940 us |       4.07% |   9.031 us |       3.65% |     0.091 us |   1.01% |   SAME   |
|   I16   |      I64      |      2^20      |  13.387 us |       2.88% |  13.392 us |       2.71% |     0.005 us |   0.03% |   SAME   |
|   I16   |      I64      |      2^24      |  67.665 us |       3.95% |  65.172 us |       3.88% |    -2.493 us |  -3.68% |   SAME   |
|   I16   |      I64      |      2^28      | 979.141 us |       1.63% | 880.944 us |       1.51% |   -98.197 us | -10.03% |   FAST   |
|   I32   |      I32      |      2^16      |   9.258 us |       4.76% |   8.894 us |       3.94% |    -0.364 us |  -3.93% |   SAME   |
|   I32   |      I32      |      2^20      |  14.452 us |       3.09% |  14.154 us |       2.46% |    -0.298 us |  -2.06% |   SAME   |
|   I32   |      I32      |      2^24      |  98.725 us |       2.56% |  95.111 us |       2.39% |    -3.613 us |  -3.66% |   FAST   |
|   I32   |      I32      |      2^28      |   1.459 ms |       0.88% |   1.406 ms |       0.87% |   -53.551 us |  -3.67% |   FAST   |
|   I32   |      I64      |      2^16      |   9.220 us |       4.55% |   8.963 us |       3.93% |    -0.256 us |  -2.78% |   SAME   |
|   I32   |      I64      |      2^20      |  14.529 us |       3.27% |  14.103 us |       2.57% |    -0.425 us |  -2.93% |   FAST   |
|   I32   |      I64      |      2^24      |  98.729 us |       2.62% |  95.279 us |       2.42% |    -3.450 us |  -3.49% |   FAST   |
|   I32   |      I64      |      2^28      |   1.459 ms |       0.92% |   1.408 ms |       0.89% |   -50.260 us |  -3.45% |   FAST   |
|   I64   |      I32      |      2^16      |   9.680 us |       4.25% |   9.349 us |       3.36% |    -0.331 us |  -3.42% |   FAST   |
|   I64   |      I32      |      2^20      |  18.766 us |       2.44% |  18.438 us |       2.28% |    -0.327 us |  -1.74% |   SAME   |
|   I64   |      I32      |      2^24      | 171.778 us |       1.28% | 171.635 us |       1.32% |    -0.143 us |  -0.08% |   SAME   |
|   I64   |      I32      |      2^28      |   2.636 ms |       0.46% |   2.638 ms |       0.50% |     1.748 us |   0.07% |   SAME   |
|   I64   |      I64      |      2^16      |   9.725 us |       4.00% |   9.336 us |       3.61% |    -0.389 us |  -4.00% |   FAST   |
|   I64   |      I64      |      2^20      |  18.757 us |       2.49% |  18.442 us |       2.15% |    -0.316 us |  -1.68% |   SAME   |
|   I64   |      I64      |      2^24      | 171.677 us |       1.31% | 171.959 us |       1.31% |     0.282 us |   0.16% |   SAME   |
|   I64   |      I64      |      2^28      |   2.636 ms |       0.46% |   2.640 ms |       0.47% |     4.692 us |   0.18% |   SAME   |
|  I128   |      I32      |      2^16      |  16.467 us |       2.84% |  16.221 us |       2.16% |    -0.246 us |  -1.49% |   SAME   |
|  I128   |      I32      |      2^20      |  64.562 us |       1.60% |  64.765 us |       1.58% |     0.203 us |   0.31% |   SAME   |
|  I128   |      I32      |      2^24      | 855.361 us |       0.40% | 874.852 us |       0.42% |    19.491 us |   2.28% |   SLOW   |
|  I128   |      I32      |      2^28      |  13.531 ms |       0.13% |  13.856 ms |       0.12% |   325.590 us |   2.41% |   SLOW   |
|  I128   |      I64      |      2^16      |  16.500 us |       2.83% |  16.354 us |       2.44% |    -0.146 us |  -0.89% |   SAME   |
|  I128   |      I64      |      2^20      |  64.526 us |       1.58% |  64.943 us |       1.56% |     0.417 us |   0.65% |   SAME   |
|  I128   |      I64      |      2^24      | 852.760 us |       0.41% | 874.305 us |       0.39% |    21.545 us |   2.53% |   SLOW   |
|  I128   |      I64      |      2^28      |  13.499 ms |       0.09% |  13.855 ms |       0.11% |   356.106 us |   2.64% |   SLOW   |
|   F32   |      I32      |      2^16      |   9.423 us |       4.13% |   9.111 us |       3.74% |    -0.312 us |  -3.31% |   SAME   |
|   F32   |      I32      |      2^20      |  14.657 us |       3.35% |  14.495 us |       3.18% |    -0.163 us |  -1.11% |   SAME   |
|   F32   |      I32      |      2^24      |  99.101 us |       2.43% |  95.987 us |       2.17% |    -3.114 us |  -3.14% |   FAST   |
|   F32   |      I32      |      2^28      |   1.461 ms |       0.88% |   1.411 ms |       0.87% |   -50.088 us |  -3.43% |   FAST   |
|   F32   |      I64      |      2^16      |   9.400 us |       4.63% |   9.095 us |       4.07% |    -0.306 us |  -3.25% |   SAME   |
|   F32   |      I64      |      2^20      |  14.775 us |       2.89% |  14.416 us |       2.12% |    -0.359 us |  -2.43% |   FAST   |
|   F32   |      I64      |      2^24      |  98.992 us |       2.35% |  96.066 us |       2.15% |    -2.926 us |  -2.96% |   FAST   |
|   F32   |      I64      |      2^28      |   1.463 ms |       0.93% |   1.412 ms |       0.87% |   -50.889 us |  -3.48% |   FAST   |
|   F64   |      I32      |      2^16      |   9.789 us |       3.13% |   9.602 us |       3.91% |    -0.187 us |  -1.91% |   SAME   |
|   F64   |      I32      |      2^20      |  18.972 us |       2.45% |  19.074 us |       2.11% |     0.102 us |   0.54% |   SAME   |
|   F64   |      I32      |      2^24      | 174.355 us |       1.38% | 171.640 us |       1.32% |    -2.715 us |  -1.56% |   FAST   |
|   F64   |      I32      |      2^28      |   2.699 ms |       0.44% |   2.641 ms |       0.47% |   -57.339 us |  -2.12% |   FAST   |
|   F64   |      I64      |      2^16      |   9.778 us |       3.91% |   9.473 us |       3.13% |    -0.304 us |  -3.11% |   SAME   |
|   F64   |      I64      |      2^20      |  18.998 us |       2.65% |  18.989 us |       2.40% |    -0.009 us |  -0.05% |   SAME   |
|   F64   |      I64      |      2^24      | 173.885 us |       1.43% | 174.820 us |       1.38% |     0.935 us |   0.54% |   SAME   |
|   F64   |      I64      |      2^28      |   2.699 ms |       0.45% |   2.703 ms |       0.47% |     4.660 us |   0.17% |   SAME   |
|   C64   |      I32      |      2^16      |  65.387 us |       1.12% |  58.121 us |       1.11% |    -7.267 us | -11.11% |   FAST   |
|   C64   |      I32      |      2^20      | 111.887 us |       1.25% | 103.188 us |       1.60% |    -8.700 us |  -7.78% |   FAST   |
|   C64   |      I32      |      2^24      |   1.105 ms |       1.50% | 938.828 us |       1.08% |  -166.071 us | -15.03% |   FAST   |
|   C64   |      I32      |      2^28      |  16.411 ms |       0.42% |  13.773 ms |       0.44% | -2638.167 us | -16.08% |   FAST   |
|   C64   |      I64      |      2^16      |  65.537 us |       1.11% |  59.344 us |       1.14% |    -6.193 us |  -9.45% |   FAST   |
|   C64   |      I64      |      2^20      | 112.330 us |       1.24% | 103.058 us |       1.29% |    -9.272 us |  -8.25% |   FAST   |
|   C64   |      I64      |      2^24      |   1.114 ms |       1.59% | 943.857 us |       1.11% |  -170.299 us | -15.29% |   FAST   |
|   C64   |      I64      |      2^28      |  16.663 ms |       0.42% |  13.787 ms |       0.39% | -2876.263 us | -17.26% |   FAST   |

# Summary

- Total Matches: 64
  - Pass    (diff <= min_noise): 32
  - Unknown (infinite noise):    0
  - Failure (diff > min_noise):  32